@@ -2,9 +2,9 @@ package com.gade.zaraproductcheckerapp.activities;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.support.v14.preference.PreferenceFragment;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.preference.PreferenceFragment;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import com.gade.zaraproductcheckerapp.R;
 import com.gade.zaraproductcheckerapp.handlers.ProductCheckerHandler;
