@@ -3,7 +3,7 @@ package com.gade.zaraproductcheckerapp.receivers;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.gade.zaraproductcheckerapp.handlers.ProductCheckerHandler;
 
