@@ -1,7 +1,7 @@
 package com.gade.zaraproductcheckerapp.util.notifications;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.gade.zaraproductcheckerapp.db.entities.ProductInfo;
 
