@@ -1,6 +1,6 @@
 package com.gade.zaraproductcheckerapp.db.converters;
 
-import android.arch.persistence.room.TypeConverter;
+import androidx.room.TypeConverter;
 
 import java.util.Date;
 
