@@ -1,16 +1,16 @@
 package com.gade.zaraproductcheckerapp.db.daos;
 
-import android.arch.persistence.room.Dao;
-import android.arch.persistence.room.Delete;
-import android.arch.persistence.room.Insert;
-import android.arch.persistence.room.Query;
-import android.arch.persistence.room.Update;
+import androidx.room.Dao;
+import androidx.room.Delete;
+import androidx.room.Insert;
+import androidx.room.Query;
+import androidx.room.Update;
 
 import com.gade.zaraproductcheckerapp.db.entities.ProductInfo;
 
 import java.util.List;
 
-import static android.arch.persistence.room.OnConflictStrategy.FAIL;
+import static androidx.room.OnConflictStrategy.FAIL;
 
 @Dao
 public interface ProductInfoDao {
